requirements.txt
asgiref==3.2.7
babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.41.0
django-localflavor==3.0.1
django-widget-tweaks==1.4.8
django==3.0.1
idna==2.9
oauthlib==3.1.0
phonenumbers==8.12.4
python-stdnum==1.13
python3-openid==3.1.0
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.23.0 ; python_version >= '3.0'
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
gunicorn==20.0.4
django-storages[google]==1.9.1